+++ src/lua/CommunityBalanceMod/Combined/Locale.lua
@@ -1,4 +1,37 @@
 if Client then
+
+    Locale.substitutions["FORTRESS_CRAG_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast healwave without a crag hive. "
+    Locale.substitutions["FORTRESS_CRAG"] = "Fortress Crag"
+    Locale.substitutions["FORTRESS_CRAG_UPGRADE"] = "Upgrade to Fortress Crag"
+    Locale.substitutions["FORTRESS_CRAG_UPGRADE_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast healwave without a crag hive."
+    Locale.substitutions["FORTRESS_CRAG_UPGRADE_RESEARCHNAME"] = "Fortress Crag"
+    Locale.substitutions["FORTRESS_CRAG_ABILITY"] = "Structual Umbra"
+    Locale.substitutions["FORTRESS_CRAG_ABILITY_TOOLTIP"] = "Cast Umbra on all nearby structures"
+   
+    Locale.substitutions["FORTRESS_SHIFT_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast echo without a shift hive. "
+    Locale.substitutions["FORTRESS_SHIFT"] = "Fortress Shift"
+    Locale.substitutions["FORTRESS_SHIFT_UPGRADE"] = "Upgrade to Fortress Shift"
+    Locale.substitutions["FORTRESS_SHIFT_UPGRADE_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast echo without a shift hive. "
+    Locale.substitutions["FORTRESS_SHIFT_UPGRADE_RESEARCHNAME"] = "Fortress Shift"
+    Locale.substitutions["FORTRESS_SHIFT_ABILITY"] = "Fortress Shift Ability"
+    Locale.substitutions["FORTRESS_SHIFT_ABILITY_TOOLTIP"] = "Fortress Shift Ability Tooltip"
+
+    Locale.substitutions["FORTRESS_SHADE_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast ink without a shade hive. "
+    Locale.substitutions["FORTRESS_SHADE"] = "Fortress Shade"
+    Locale.substitutions["FORTRESS_SHADE_UPGRADE"] = "Upgrade to Fortress Shade"
+    Locale.substitutions["FORTRESS_SHADE_UPGRADE_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast ink without a shade hive. "
+    Locale.substitutions["FORTRESS_SHADE_UPGRADE_RESEARCHNAME"] = "Fortress Shade"
+    Locale.substitutions["FORTRESS_SHADE_ABILITY"] = "Fortress Shade Ability"
+    Locale.substitutions["FORTRESS_SHADE_ABILITY_TOOLTIP"] = "Create 5 hallucination structures"
+
+    Locale.substitutions["FORTRESS_WHIP_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast ABILITY without a whip hive. "
+    Locale.substitutions["FORTRESS_WHIP"] = "Fortress Whip"
+    Locale.substitutions["FORTRESS_WHIP_UPGRADE"] = "Upgrade to Fortress Whip"
+    Locale.substitutions["FORTRESS_WHIP_UPGRADE_TOOLTIP"] = "Triple the HP, moves significant slower, able to cast ABILITY without a whip hive. "
+    Locale.substitutions["FORTRESS_WHIP_UPGRADE_RESEARCHNAME"] = "Fortress Whip"
+    Locale.substitutions["FORTRESS_WHIP_ABILITY"] = "Fortress Whip Ability"
+    Locale.substitutions["FORTRESS_WHIP_ABILITY_TOOLTIP"] = "Fortress Whip Ability Tooltip"
+
     Locale.substitutions["RESILIENCE"] = "Heat Plating"
     Locale.substitutions["RESILIENCE_TOOLTIP"] = "Reduced damage from explosive, fire, and energy based weaponry"
     Locale.substitutions["ADVANCED_SUPPORT"] = "Advanced Support"