+++ src/lua/CommunityBalanceMod/Combined/TeamInfo.lua
@@ -1,0 +1,94 @@
+
+
+-- Relevant techs must be ordered with children techs coming after their parents
+TeamInfo.kRelevantTechIdsMarine =
+{
+
+    kTechId.ShotgunTech,
+    --kTechId.HeavyMachineGunTech,
+    kTechId.MinesTech,
+    kTechId.WelderTech,
+    kTechId.GrenadeTech,
+    
+    kTechId.AdvancedArmory,
+    kTechId.AdvancedArmoryUpgrade,
+    kTechId.AdvancedWeaponry,
+
+    kTechId.Weapons1,
+    kTechId.Weapons2,
+    kTechId.Weapons3,
+    kTechId.Armor1,
+    kTechId.Armor2,
+    kTechId.Armor3,
+
+    kTechId.PrototypeLab,
+    kTechId.UpgradeToAdvancedPrototypeLab, -- CommunityBalanceMod 
+    kTechId.JetpackTech,
+    kTechId.ExosuitTech,
+    kTechId.DualMinigunTech,
+
+    kTechId.ARCRoboticsFactory,
+    kTechId.UpgradeRoboticsFactory,
+    kTechId.MACEMPTech,
+    kTechId.MACSpeedTech,
+    
+    kTechId.Observatory,
+    kTechId.PhaseTech,
+
+    kTechId.AdvancedMarineSupport,
+}
+
+
+--FortressPvE
+TeamInfo.kRelevantTechIdsAlien =
+{
+
+    kTechId.CragHive,
+    kTechId.UpgradeToCragHive,
+    kTechId.Shell,
+    kTechId.TwoShells,
+    kTechId.ThreeShells,
+    
+    kTechId.ShadeHive,
+    kTechId.UpgradeToShadeHive,
+    kTechId.Veil,
+    kTechId.TwoVeils,
+    kTechId.ThreeVeils,
+    
+    kTechId.ShiftHive,
+    kTechId.UpgradeToShiftHive,
+    kTechId.Spur,
+    kTechId.TwoSpurs,
+    kTechId.ThreeSpurs,
+
+    -- CommunityBalanceMod
+    -- limited to 32 due to bitmasking overflow, see utility: CreateBitMask
+    --[[
+   kTechId.FortressCrag,
+    kTechId.UpgradeToFortressCrag,
+    kTechId.FortressShift,
+    kTechId.UpgradeToFortressShift,
+    kTechId.FortressShade,
+   kTechId.UpgradeToFortressShade,
+     kTechId.FortressWhip,
+    kTechId.UpgradeToFortressWhip,]]
+
+    
+    kTechId.ResearchBioMassOne,
+    kTechId.ResearchBioMassTwo,
+    kTechId.ResearchBioMassThree,
+
+    kTechId.Leap,
+    kTechId.Xenocide,
+    kTechId.BileBomb,
+    kTechId.WebTech,
+    kTechId.Umbra,
+    kTechId.Spores,
+    kTechId.MetabolizeEnergy,
+    kTechId.MetabolizeHealth,
+    kTechId.Stab,
+    kTechId.Charge,
+    kTechId.BoneShield,
+    kTechId.Stomp,
+    
+}