+++ src/lua/CommunityBalanceMod/Combined/TechData.lua
@@ -40,6 +40,178 @@ local function GetTechToAdd()
             [kTechDataTooltipInfo] = "EXOSUIT_TECH_TOOLTIP", -- "Allows Exosuits to be purchased"   text for desciption
             [kTechDataResearchName] = "RESEARCH_EXOSUITS_TITLE" -- "Exosuits"               text for left side
         },
+
+        {
+            [kTechDataId] = kTechId.FortressCrag,
+            [kTechDataBioMass] = kCragBiomass,
+            [kTechDataSupply] = kCragSupply,
+            [kTechDataGhostModelClass] = "AlienGhostModel", 
+            [kTechDataMapName] = FortressCrag.kMapName, 
+            [kTechDataDisplayName] = "FORTRESS_CRAG", 
+            [kTechDataRequiresInfestation] = true, 
+            [kTechDataModel] = Crag.kModelName, 
+            [kTechDataMaxHealth] = kFortressCragHealth,
+            [kTechDataMaxArmor] = kFortressCragArmor,
+            [kTechDataInitialEnergy] = kCragInitialEnergy,
+            [kTechDataMaxEnergy] = kCragMaxEnergy, 
+            [kTechDataPointValue] = kCragPointValue * 2,
+            [kVisualRange] = Crag.kHealRadius, 
+            [kTechDataTooltipInfo] = "FORTRESS_CRAG_TOOLTIP", 
+            [kTechDataGrows] = false, 
+            [kTechDataObstacleRadius] = 1.15, 
+        },
+        {
+            
+            
+            [kTechDataId] = kTechId.UpgradeToFortressCrag,
+            [kTechDataCostKey] = kFortressUpgradeCost,
+            [kTechIDShowEnables] = false,
+            [kTechDataResearchTimeKey] = kFortressResearchTime,
+            [kTechDataDisplayName] = "FORTRESS_CRAG_UPGRADE",
+            [kTechDataTooltipInfo] = "FORTRESS_CRAG_UPGRADE_TOOLTIP",
+            [kTechDataResearchName] = "FORTRESS_CRAG_UPGRADE_RESEARCHNAME",
+            [kTechDataOneAtATime] = true,
+        },
+
+        {
+            [kTechDataId] = kTechId.FortressCragAbility,
+            [kTechDataCooldown] = kFortressAbilityCooldown, 
+            [kTechDataDisplayName] = "FORTRESS_CRAG_ABILITY", 
+            [kTechDataCostKey] = kFortressAbilityCost,
+            [kTechDataTooltipInfo] = "FORTRESS_CRAG_ABILITY_TOOLTIP", 
+            [kTechDataOneAtATime] = true,  
+        },
+
+        {
+            [kTechDataId] = kTechId.FortressShift,
+            [kTechDataBioMass] = kShiftBiomass,
+            [kTechDataSupply] = kShiftSupply,
+            [kTechDataGhostModelClass] = "ShiftGhostModel", 
+            [kTechDataMapName] = FortressShift.kMapName, 
+            [kTechDataDisplayName] = "FORTRESS_SHIFT", 
+            [kTechDataRequiresInfestation] = true, 
+            [kTechDataModel] = Shift.kModelName, 
+            [kTechDataMaxHealth] = kFortressShiftHealth,
+            [kTechDataMaxArmor] = kFortressShiftArmor, 
+            [kTechDataInitialEnergy] = kShiftInitialEnergy, 
+            [kTechDataMaxEnergy] = kShiftMaxEnergy, 
+            [kTechDataPointValue] = kShiftPointValue * 2,
+            [kVisualRange] =  {
+                kEchoRange,
+                kEnergizeRange
+            },
+            [kTechDataTooltipInfo] = "FORTRESS_SHIFT_TOOLTIP", 
+            [kTechDataGrows] = false, 
+            [kTechDataObstacleRadius] = 1.3,
+        },
+        {
+            
+            
+            [kTechDataId] = kTechId.UpgradeToFortressShift,
+            [kTechDataCostKey] = kFortressUpgradeCost,
+            [kTechIDShowEnables] = false,
+            [kTechDataResearchTimeKey] = kFortressResearchTime,
+            [kTechDataDisplayName] = "FORTRESS_SHIFT_UPGRADE",
+            [kTechDataTooltipInfo] = "FORTRESS_SHIFT_UPGRADE_TOOLTIP",
+            [kTechDataResearchName] = "FORTRESS_SHIFT_UPGRADE_RESEARCHNAME",
+            [kTechDataOneAtATime] = true,
+        },
+
+        {
+            [kTechDataId] = kTechId.FortressShiftAbility,
+            [kTechDataCooldown] = kFortressAbilityCooldown, 
+            [kTechDataDisplayName] = "FORTRESS_SHIFT_ABILITY", 
+            [kTechDataCostKey] = kFortressAbilityCost,
+            [kTechDataTooltipInfo] = "FORTRESS_SHIFT_ABILITY_TOOLTIP", 
+            [kTechDataOneAtATime] = true,  
+        },
+
+            
+        {
+            [kTechDataId] = kTechId.FortressShade,
+            [kTechDataBioMass] = kShadeBiomass,
+            [kTechDataSupply] = kShadeSupply,
+            [kTechDataGhostModelClass] = "AlienGhostModel",
+            [kTechDataMapName] = FortressShade.kMapName, 
+            [kTechDataDisplayName] = "FORTRESS_SHADE", 
+            [kTechDataRequiresInfestation] = true, 
+            [kTechDataModel] = Shade.kModelName, 
+            [kTechDataMaxHealth] = kFortressShadeHealth,
+            [kTechDataMaxArmor] = kFortressShadeArmor, 
+            [kTechDataInitialEnergy] = kShadeInitialEnergy, 
+            [kTechDataMaxEnergy] = kShadeMaxEnergy, 
+            [kTechDataPointValue] = kShadePointValue * 2,
+            [kVisualRange] =  Shade.kCloakRadius,
+            [kTechDataTooltipInfo] = "FORTRESS_SHADE_TOOLTIP", 
+            [kTechDataGrows] = false, 
+            [kTechDataObstacleRadius] = 1.25,
+            [kTechDataMaxExtents] = Vector(1, 1.3, .4),
+        },
+        {
+            
+            
+            [kTechDataId] = kTechId.UpgradeToFortressShade,
+            [kTechDataCostKey] = kFortressUpgradeCost,
+            [kTechIDShowEnables] = false,
+            [kTechDataResearchTimeKey] = kFortressResearchTime,
+            [kTechDataDisplayName] = "FORTRESS_SHADE_UPGRADE",
+            [kTechDataTooltipInfo] = "FORTRESS_SHADE_UPGRADE_TOOLTIP",
+            [kTechDataResearchName] = "FORTRESS_SHADE_UPGRADE_RESEARCHNAME",
+            [kTechDataOneAtATime] = true, 
+        },
+
+        {
+            [kTechDataId] = kTechId.ShadeHallucination,
+            [kTechDataCooldown] = kFortressAbilityCooldown, 
+            [kTechDataDisplayName] = "FORTRESS_SHADE_ABILITY", 
+            [kTechDataCostKey] = kFortressAbilityCost,
+            [kTechDataTooltipInfo] = "FORTRESS_SHADE_ABILITY_TOOLTIP", 
+            [kTechDataOneAtATime] = true,  
+        },
+            
+        {
+            [kTechDataId] = kTechId.FortressWhip,
+            [kTechDataBioMass] = kWhipBiomass,
+            [kTechDataSupply] = kWhipSupply,
+            [kTechDataGhostModelClass] = "AlienGhostModel",
+            [kTechDataMapName] = FortressWhip.kMapName,  
+            [kTechDataDisplayName] = "FORTRESS_WHIP", 
+            [kTechDataRequiresInfestation] = true, 
+            [kTechDataModel] = Whip.kModelName, 
+            [kTechDataMaxHealth] = kFortressWhipHealth,
+            [kTechDataMaxArmor] = kFortressWhipArmor, 
+            [kTechDataInitialEnergy] = kWhipInitialEnergy,  
+            [kTechDataDamageType] = kDamageType.Structural,
+            [kTechDataMaxEnergy] = kWhipMaxEnergy, 
+            [kTechDataPointValue] = kWhipPointValue * 2,
+            [kVisualRange] = Whip.kRange,
+            [kTechDataTooltipInfo] = "FORTRESS_WHIP_TOOLTIP", 
+            [kTechDataGrows] = false, 
+            [kTechDataObstacleRadius] = 0.85,
+       
+        },
+        {
+            
+            
+            [kTechDataId] = kTechId.UpgradeToFortressWhip,
+            [kTechDataCostKey] = kFortressUpgradeCost,
+            [kTechIDShowEnables] = false,
+            [kTechDataResearchTimeKey] = kFortressResearchTime,
+            [kTechDataDisplayName] = "FORTRESS_WHIP_UPGRADE",
+            [kTechDataTooltipInfo] = "FORTRESS_WHIP_UPGRADE_TOOLTIP",
+            [kTechDataResearchName] = "FORTRESS_WHIP_UPGRADE_RESEARCHNAME",
+            [kTechDataOneAtATime] = true, 
+        },
+
+        {
+            [kTechDataId] = kTechId.FortressWhipAbility,
+            [kTechDataCooldown] = kFortressAbilityCooldown, 
+            [kTechDataDisplayName] = "FORTRESS_WHIP_ABILITY", 
+            [kTechDataCostKey] = kFortressAbilityCost,
+            [kTechDataTooltipInfo] = "FORTRESS_WHIP_ABILITY_TOOLTIP", 
+            [kTechDataOneAtATime] = true,  
+        },
+
     }
 end
 
@@ -80,7 +252,7 @@ local function GetTechToChange()
         },
 
         -- In vanilla EN its called "Advanced Support" and in every other language except russian "Advanced Assistance"
-        -- this will overwrite the russian translation
+        -- this will overwrite the russian translation too
         [kTechId.AdvancedMarineSupport] = 
         {
             [kTechDataDisplayName] = "ADVANCED_SUPPORT",