+++ src/lua/CommunityBalanceMod/FortressPvE/TwiliteShade/TechTreeButtons.lua
@@ -1,0 +1,8 @@
+local kTechIdToMaterialOffset = debug.getupvaluex(GetMaterialXYOffset, "kTechIdToMaterialOffset")
+
+kTechIdToMaterialOffset[kTechId.ShadeHallucination] = 126
+
+kTechIdToMaterialOffset[kTechId.HallucinateShell] = 22
+kTechIdToMaterialOffset[kTechId.HallucinateSpur] = 11
+kTechIdToMaterialOffset[kTechId.HallucinateVeil] = 23
+kTechIdToMaterialOffset[kTechId.HallucinateEgg] = 34